requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Werkzeug==2.3.7
wtforms==3.0.1
scikit-learn==1.5.0
pandas==2.2.3
numpy==1.26.4
gunicorn==21.2.0
